Se modificaron arhivos

+++ scripts/before_install.bat
@@ -1,3 +1,3 @@
-@echo off
-rmdir /s /q c:\inetpub\wwwroot
+REM Limpiar el directorio de destino
+if exist c:\inetpub\wwwroot rmdir /s /q c:\inetpub\wwwroot
 mkdir c:\inetpub\wwwroot
